fix: should check if is_enabled() already exists before defining it
README sensibly suggests adding this block to your theme so that calls to `is_enabled()` will not break your theme if the plugin is disabled.

```
if ( ! function_exists( 'is_enabled' ) ) {
	function is_enabled() {
		return false;
	}
}
```
However, if you add this block to `functions.php` file, then try to activate the plugin, you get an error 500 in WordPress. This patch fixes that.